Refactor party_calculate_loot script Changes how the plunder (merchandise) amount is calculated.  The formula now looks at the type of party and adds trade goods or food.  Caravans and bandit lairs will now start dropping merchandise in their loot piles again.
The maximum amount of plunder that you can get
is limited by the party/player looting skill.

Additional change - the chance of an item dropping is
no longer (100,000 / price) but (100,000 / sqrt(price)).
This should help useful items for the player to drop
without being crowded out by sticks and stones.  It
might be too powerful of a change.